Add basic pages controller
This controller still has no authentication or authorization and will
show you any page. BUT it will send you all the things you wish for.

This can not be deployed on a live site yet, as it will also show you
all the things related to the site if you ask the right way.