Add IDs to Alchemy JSON serializers
JSON:API returns a top-level ID for each serialized object, however
this ID is always a string and does not necessarily correspond to a
database ID. It's an internal ID used to match objects within the
JSON response. Therefore, we should be including the actual database
ID as an attribute in our serializers. This has the added benefit of
ensuring the ID is returned as an integer in the JSON response.

https://jsonapi.org/format/#document-resource-object-identification