Rename and expose Page#page_version_type
The former named argument was misleading as it did not reflected
the page version, but a symbol of the type of version we want
to use from the Alchemy::Page.

As we want to expose the page_version_type, so we can use it and
pass on to new instances we needed to clarify.