Add srcset support to IngredientPicture serializer
If srcset is defined on the element definition, then it returns an array of
image candidates its descriptor and the image dimensions.

eg:

```json
{
  "url": "cute-kittins.jpg",
  "desc": "200w",
  "width": "200",
  "height": "100",
}
```

You can define srcset candidates either as a size String or a Hash
supporting all image transformations Alchemy supports. So you are
able to create cropped jpgs from a png for instance.