Rely on improved etag for caching
The Last-Modified header is a weaker cache key than the etag, and thus
we rely solely on the etag from now on. Browsers and Rails will ignore
it if it's not set while validating if a request is fresh.

Also sets the etag based on JSONAPI's typical set of params, so that we
get a different etag for different `include`, `sort`, `page`, `fields`,
or `filter` params.