chore: replace rollup with vite
In order to make development easier and stay in a ecosystem more related to typescript we replace rollup with vite.
It is common pattern to give the responsibility of transpiling js to support older version to the production implementation of the module. We follow that pattern and remove babel from the project

THIS IS A BREAKING CANGE: This repo used to transpile the js with babel. That is now the responsibility of the module implementation