feat: remove deprecation filtering
We remove the filtering of deprecated elements.

If an deprecated element should be displayed or not is highly dependent on the implementation and should be decided in the backend or implementation of the data

THIS IS A BREAKING CHANGE: deserializePage and deserializePages have been removed

If you relied on this module removing deprecated elements from your api response you need to re-implement that custom.