Lower cache duration to 10 minutes
max-gae means the browser or CDN waits until this time passed to revalidate the stale cache on the origin server.

Since we use an etag to match the cache key, ther server will
response with 304 (not modified) if the content is still the same. that way we can ensure that new content will be available in 10 minutes after publishing the page.

Combined with `stale-while-revalidate` this provides a very efficient caching strategy.